+++ lib/generate-footprint-tsx/convert-copper-text.ts
@@ -1,6 +1,6 @@
 import { su } from "@tscircuit/soup-util"
 import type { FootprintElementConverter } from "./converter-types"
-import { escapeJsxText } from "./footprint-tsx-attribute-formatters"
+import { escapeJsxText } from "./footprint-tsx-attribute-formatters/escape-jsx-text"
 
 export const convertCopperText: FootprintElementConverter = (circuitJson) => {
   const copperTexts = su(circuitJson).pcb_copper_text.list()
