+++ lib/generate-footprint-tsx/convert-silkscreen-text.ts
@@ -1,7 +1,7 @@
 import { su } from "@tscircuit/soup-util"
 import { mmStr } from "@tscircuit/mm"
 import type { FootprintElementConverter } from "./converter-types"
-import { escapeJsxText } from "./footprint-tsx-attribute-formatters"
+import { escapeJsxText } from "./footprint-tsx-attribute-formatters/escape-jsx-text"
 
 export const convertSilkscreenText: FootprintElementConverter = (
   circuitJson,
