+++ lib/generate-footprint-tsx/footprint-tsx-attribute-formatters/escape-jsx-text.ts
@@ -1,0 +1,2 @@
+export const escapeJsxText = (value: unknown): string =>
+  String(value ?? "").replace(/"/g, '\\"')
