+++ lib/generate-footprint-tsx/footprint-tsx-attribute-formatters/format-mm.ts
@@ -1,0 +1,3 @@
+import { mmStr } from "@tscircuit/mm"
+
+export const formatMm = (value: number | undefined): string => mmStr(value ?? 0)
