+++ lib/generate-footprint-tsx/footprint-tsx-attribute-formatters/format-optional-mm-attr.ts
@@ -1,0 +1,9 @@
+import { mmStr } from "@tscircuit/mm"
+
+export const formatOptionalMmAttr = (
+  attrName: string,
+  value: number | undefined,
+): string => {
+  if (value === undefined) return ""
+  return ` ${attrName}="${mmStr(value)}"`
+}
