+++ lib/generate-footprint-tsx/footprint-tsx-attribute-formatters/format-pcb-rotation-attr.ts
@@ -1,0 +1,8 @@
+export const formatPcbRotationAttr = (
+  rotation: string | number | undefined,
+  attrName = "pcbRotation",
+): string => {
+  if (rotation === undefined) return ""
+  const formatted = typeof rotation === "number" ? `${rotation}deg` : rotation
+  return ` ${attrName}="${formatted}"`
+}
